Delete old DiskImage download code
Remove diskimage.go (708 lines) and diskimage_test.go (1,242 lines)
which are fully replaced by bootmedia.go. Move shared constants
(downloadRequestTimeout, checksumDiscoveryTimeout) and parseChecksumFile
function to bootmedia.go. Move fakeHTTPDoer test helper to
testhelpers_test.go. Remove dead isoBasePath field and SetISOBasePath
method, replace with filesBasePath/--files-path flag in main.go.
