+++ cmd/isoboot-controller/main.go
@@ -19,12 +19,12 @@ func main() {
 	var (
 		port        string
 		namespace   string
-		isoBasePath string
+		filesBasePath string
 	)
 
 	flag.StringVar(&port, "port", "8081", "gRPC server port")
 	flag.StringVar(&namespace, "namespace", "", "Kubernetes namespace")
-	flag.StringVar(&isoBasePath, "iso-path", "/opt/isoboot/iso", "Base path for ISO storage")
+	flag.StringVar(&filesBasePath, "files-path", "/opt/isoboot/files", "Base path for boot file storage")
 	flag.Parse()
 
 	if namespace == "" {
@@ -47,9 +47,9 @@ func main() {
 	}
 
 	// Create and start controller
-	// NOTE: SetISOBasePath must be called before Start.
+	// NOTE: SetFilesBasePath must be called before Start.
 	ctrl := controller.New(k8sClient)
-	ctrl.SetISOBasePath(isoBasePath)
+	ctrl.SetFilesBasePath(filesBasePath)
 	ctrl.Start()
 	defer ctrl.Stop()
