+++ internal/controller/testhelpers_test.go
@@ -1,6 +1,8 @@
 package controller
 
 import (
+	"net/http"
+
 	"github.com/isoboot/isoboot/internal/k8s/typed"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -31,6 +33,15 @@ func newTypedConfigMap(name string, data map[string]string) *corev1.ConfigMap {
 		ObjectMeta: metav1.ObjectMeta{Name: name, Namespace: "default"},
 		Data:       data,
 	}
+}
+
+// fakeHTTPDoer is a mock HTTP client for tests
+type fakeHTTPDoer struct {
+	doFunc func(req *http.Request) (*http.Response, error)
+}
+
+func (f *fakeHTTPDoer) Do(req *http.Request) (*http.Response, error) {
+	return f.doFunc(req)
 }
 
 // newTypedSecret is a helper to create a corev1.Secret for testing
